requirements-lambda.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
strictyaml>=1.0.0
